+++ src/main/kotlin/cn/varsa/idea/pde/partial/common/Constants.kt
@@ -22,9 +22,6 @@ const val ProjectLibraryNamePrefix = PartialPrefix
 const val KotlinBundleSymbolName = "org.jetbrains.kotlin.osgi-bundle"
 const val KotlinOrderEntryName = "KotlinJavaRuntime"
 
-const val KILOBYTE = 1024
-const val MEGABYTE = KILOBYTE * KILOBYTE
-
 const val Java = "java"
 const val Kotlin = "kotlin"